add export data in view to advanced search (and reorganize utils code a bit)
